requirements.txt
paddlepaddle<=2.6
paddleocr>=2.7,<3
Pillow>=10,<12
